requirements-dev.txt
-r requirements.txt
pytest
pytest-django
coverage
flake8
mypy
bandit